changed division operation to use float instead of integer

+++ Simple Calculation Software/Header.h
@@ -15,6 +15,6 @@
 int addFn(int first, int second);
 int subFn(int first, int second);
 int mulFn(int first, int second);
-int divFn(int first, int second);
+float divFn(int first, int second);
 
 #endif /* Header_h */
